Move auth header into reusable function
