+++ src/services/config.js
@@ -1,27 +1,8 @@
 import ky from 'ky'
-import { CognitoUserPool } from 'amazon-cognito-identity-js'
 
 // FPCC
 import GlobalConfiguration from 'src/GlobalConfiguration'
-
-const poolData = {
-  UserPoolId: GlobalConfiguration.AWS_USER_POOL_ID,
-  ClientId: GlobalConfiguration.AWS_CLIENT_ID,
-}
-const userPool = new CognitoUserPool(poolData)
-const cognitoUser = userPool.getCurrentUser()
-
-let headers = {}
-
-if (cognitoUser != null) {
-  cognitoUser.getSession((err, session) => {
-    if (session?.isValid()) {
-      headers = {
-        Authorization: `Bearer ${session.getAccessToken().getJwtToken()}`,
-      }
-    }
-  })
-}
+import { getAuthHeaderIfTokenExists } from 'common/utils/authHelpers'
 
 export const apiV1 = ky.create({
   prefixUrl: GlobalConfiguration.V1_API_URL,
@@ -31,7 +12,7 @@ export const apiV1 = ky.create({
 export const apiBase = ky.create({
   prefixUrl: GlobalConfiguration.API_URL,
   timeout: 60000,
-  headers,
+  headers: getAuthHeaderIfTokenExists(),
 })
 
 export const externalApi = ky.create({
